Refactor paths; Move MultiClient and Client to clients.py; Move base clients part to baseclients.py;
